LF6 update (#368)
* Update 06.ptx

Answer choices A and C were the same (6c + 4a vs 4a + 6c). The answer said A, so I assume C was supposed to be different and switched the variables to make what I think was the intended wrong answer.

* Update 06.ptx

made all the variables in the same order

* Update 06.ptx

fixed other question right before that also had inconsistent order

* added in dollar signs

---------

Co-authored-by: Steven Clontz <steven.clontz@gmail.com>