m-and-ns in EV2
The main fact in EV2 has the number of vectors as n as well as the dimension of the ambient space. I have changed it so that it about n vectors in R^m and I chose this convention so that it's an mxn matrix that we ultimately consider. 

Given this change, I also changed the n to an m in EV2-n-vectors-span (just to not throw folks off). I didn't change the label of this fact though so that we don't have to search for where it's used.